A JvmStats Class

An S4 class to represent statistics related to the Java Virtual Machine (JVM).

Displays JVM statistics in a formatted manner.

Usage

# S4 method for class 'JvmStats'
show(object)

Arguments

object

An object of class JvmStats.

Value

An object of class JvmStats.

Details

This method provides a structured and formatted output of the JVM statistics. It displays the number of JARs, the Java version, the PHG version, and detailed memory statistics (maximum memory, total memory, free memory, and allocated memory) in a human-readable format. Memory values are shown in the respective units (e.g., bytes, MB, GB) defined in the object.

Slots

nJars

An integer representing the number of JAR files currently loaded in the JVM.

javaVersion

A character string representing the version of Java being used.

phgVersion

A character string representing the version of PHG (Practical Haplotype Graph).

classPath

A character vector representing the JAR files found in the class path.

maxMem

A numeric value representing the maximum amount of memory (in bytes) that the JVM can use.

totMem

A numeric value representing the total amount of memory (in bytes) currently allocated to the JVM.

freeMem

A numeric value representing the amount of free memory (in bytes) in the JVM.

allocMem

A numeric value representing the amount of memory (in bytes) currently allocated by the JVM.

memUnit

A character string representing the unit of memory (e.g., "bytes", "MB", "GB").