Generate GVCF plots

Generates general GVCF metrics plots for genome-wide GVCF statistics

Usage

plotGvcf(object, ...)

# S4 method for class 'PHGMetrics'
plotGvcf(
  object,
  metricId = NULL,
  f = NULL,
  nRow = NULL,
  nCol = NULL,
  tag = "A",
  mData = NULL,
  mVar = NULL
)

Arguments

object

A PHGMetrics object containing the gVCF data.

...

Additional arguments, for use in specific methods

metricId

A character vector specifying the ID of the metric to be plotted. Only a single ID is supported.

f

A formula object defining the plot.

nRow

An integer specifying the number of rows in the plot layout.

nCol

An integer specifying the number of columns in the plot layout.

tag

What tag type do you want passed to final plot?

mData

Add optional metadata data.frame object for categorical plotting. This object must contain a singular column named sample, taxa, or line along with one or more columns containing categorical data for plotting.

mVar

If mData is specified, what categorical column do you want plotted? If NULL, the first non-sample/taxa/line column will be selected.

Value

A plot object generated from the specified gVCF data and layout.