Return DNA sequence data

Gets DNA sequence from a PHGDataSet

This S4 method reads sequence data from a PHGDataSet object. It is linked to the AGC CLI program for sequence retrieval. The method takes optional parameters to specify read range IDs or haplotype IDs, and an optional pad value.

Usage

readSequence(object, ...)

# S4 method for class 'PHGDataSet'
readSequence(object, rrId = NULL, hapId = NULL, pad = 0)

Arguments

object

A PHGDataSet object.

...

Additional arguments, for use in specific methods

rrId

An reference range ID. Defaults to NULL. If specified, this will return all haplotype sequences for each sample in a given reference range.

hapId

A haplotype ID. Defaults to NULL.

pad

An integer value for padding around the sequence region. Defaults to 0.

Value

Returns sequence data corresponding to the specified parameters.

Details

This method calls readSequenceFromPds() to retrieve the sequence data. Additional arguments may be used to refine the range.

Examples

if (FALSE) { # \dontrun{
myPds <- PHGDataSet()
readSequence(myPds, rrId = 1, hapId = "hapA", pad = 100)
} # }